Limit stdio.h inclusion before sys-core.h
The inclusion of stdio.h is avoided in the core, to keep from pulling
in `printf` and `fprintf` and related routines.  There are some hacks
in %sys-core.h that help release builds avoid these inclusions, but
it didn't catch some inclusions of stdio.h *before* %sys-core.h

This adds another check to prevent that.  But there has to be a
loophole on POSIX because the file API for `rename()` is only available
in stdio.h.  A special define is used to guard this condition.