Eliminate stdio.h dependencies in Windows release builds
The added check that stdio was not being included in release builds
pointed out cases in Windows where it was being included.  This
resolves that, as well as tidying up other dependencies in the
Crypt repository on snprintf(), permitting the paring out of the
`oid.c` file.

This process discovered an unfortunate aspect, that including the
C++ `<string>` library would pull in stdio.h on Windows.  So C++
builds on Windows would either have to not offer the ability to turn
std::string into Rebol values via libRebol, or disable the check
that stdio.h is not included.