Loosen rules for irregular sizes
In order to make sure that memory balanced out to an even
amount at the end of the run, there was a point in time when
it was needed to make allocations for series elements be
a precise power of 2.  This does not appear to be a current
requirement (?) so the assert is commented out, until such
time that it is certain to not be needed.