Adjust uses of "Ensure_XXX" in C sources
The first uses of the word "Ensure" in the C code were introduced on
functions that would pass through the input value, and apply a needed
operation on the value if it was needed.  These were routines that
were designed *not* to raise errors--but to "ensure" the input was
passed through with the proper procedure applied (if needed).

This is a different sense of "ensure" from what arose as natives, e.g.
ENSURE is ensuring that the program does not continue if the condition
is not true...but passes through the input value.

What both of these senses of ensure have in common is that the value
is passed through unmodified if it's already "ok".  But beyond that,
the behavior is very different.

This commit changes the mutating version to "Force_XXX".  The prefix
of "force" helps convey the notion that the code intends to keep
going one way or another.

It also changes some other odds and ends of usage of Ensure to other
words as appropriate.  This helps ENSURE stay in the category of raising
an error if the condition does not hold...like preconditions in Eiffel:

https://archive.eiffel.com/doc/online/eiffel50/intro/language/invitation-07.html