Get rid of SER_SEEK
SER_SEEK is an artifact from the early stages of designing series that
have a variable-sized conception of "length", but a fixed conception of
their element size (width).

Ultimately there was only one series type that chose to work this way:
UTF-8 series stored in REBSTR*.  The mechanics for dealing with it
wound up being managed at a higher level than REBSER.  So SER_SEEK
ended up being pared down into a total synonym for SER_AT.

This cleans up the vestigial function.