Friendlier name SER_XXX_RAW => SER_DATA_XXX
Ren-C drastically cut down on errors from using the wrong series
by having compile-time and run-time checks matching the size of the
units in the series with the callsite.  Higher level accessors like
SER_AT() would be macros parameterized with a data type, and the size
of that data type would be passed a the lower-level SER_AT_RAW()
routine that would assert the size of the type matched the runtime
size of the series.  The macro would also cast the REBYTE* raw byte
location into the type passed in.

This works well.  However, the "_RAW" name is a little intimidating.
Now that the codebase almost never uses the raw form, the cases that
do stand out enough on their own by virtue of requiring you to take
the unusual step of passing in an element size...which is enough to
cause a caller to pause and ask if that's what they really want to
do.  Also, the SER_DATA_RAW() function which was effectively
SER_HEAD_RAW() is actually useful...and can be changed to the simple
name SER_DATA().