Rename KNOWN to SPECIFIC
The years-ago specific binding transition introduced a macro to be used
as a convenience in cases where a value was "known" to not be the type
of value that had a binding which had to be taken into consideration.
Debug builds would assert this was actually the case...making it better
than having the callsites do an unchecked `cast(REBVAL*, some_relval)`.

But the word KNOWN() can apply more generally to things that you know,
like KNOWN_MUTABLE().  This could be changed to KNOWN_SPECIFIC(), but
that would be rather wordy for this very common function.  Use the
shorter form of SPECIFIC().