Get rid of "round flags" (RF_XXX)
R3-Alpha factored the implementation of rounding into some service
routines that were called from different ACTION! implementations of
the ROUND native code.  This took a number of flags, but there was
no way to pass around named arguments in a frame.

Consequently, there was some busywork of unpacking the nice refinements
from the frame into bitflags, and having to keep that in alignment.
With Ren-C there are frames and macros for talking about the arguments
by name in a way that stays in sync.  So these service routines that
are only called from the ROUND natives can just pass the frame.

It saves on the busywork, shortens the code, uses less memory, and is
overall faster.

Note: It turned out there was one case in tuple division that used
the round function from a divide operation.  Hence no ROUND frame was
available.  These cases are best addressed by building a round frame
and calling round, vs. trying to make a parallel invocation structure
that mimics a frame...then try and optimize *that* so it benefits the
system across the board.  The tuple code is all going to be replaced,
in any event...