Add FROZEN_SHALLOW concept, tidy Frozen manipulation
The concept of "deep freeze" of a series was added to enable MAP! to
have blocks and strings as keys.  Whatever would be used to hash them
could then guarantee that the hash was still valid when it went to
search for the thing later.  (This was based on the idea that being
able to search for strings and blocks as keys in MAP!s was not
negotiable, and that making copies was wasteful if the original data
was not needed for other purposes.)

With PATH! existing as a type that enforces top-level mutability, this
introduces the need for shallow freezing.  This allows the paths to
be aliased AS BLOCK!, and also is a general sanity check to make sure
that paths aren't getting modified on accident.