Temporarily reclaim reskin parameter bit
Reskinning a function can mean that even though functional compositions
think they've type checked something, the skin might have allowed
incorrect types through and need to be rechecked.

To save on duplicate checking, the skinning process was using up a bit
in the typeset to say when the checking was needed.  Typeset bits are
in short supply right now and the duplicate checking is not that slow
in its current form, so this just rechecks all the arguments on
a type expansion.