Disable REFINEMENT! parameter type checking
Making REFINEMENT!, `/`, `//`, etc. instances of a more clean and
powerful path type has been difficult, but is solidifying as a far
superior concept.  However, type checking and matching of these sub
patterns in function prototypes is not supported.

A trick was used to make it so that REFINEMENT! could still be matched
as a parse, by special coding:

    refinement!: @refinement!

So when PARSE recognizes that special symbol it knows what to do.
However, a parallel hack in the function parameter type checking needs
to store a typeset bit somewhere...and these bits have run out and
need to be reclaimed for a TUPLE! implementation.

This puts the responsibility for sub-checking PATH! forms upon the
function implementer for now...though implementation of constraints
and predicates is being prioritized in the implementation of TUPLE!