Reorder types to make TUPLE! bindable
The packed form of tuple would store its length in the EXTRA field.
Bindable types use EXTRA for the binding, so with TUPLE! shifting to
share implementation code with PATH! this needed to change.

In addition, this brings about stricter enforcement of the immutability
of tuples, by forcing creation of new tuples instead of fiddling with
the bits of existing ones.