Change path dispatch to not see picker binding
Originally it was conceived that a REBVAL* had to be passed to eac
path processing step, as opposed to a RELVAL* (incomplete binding,
inspect by structure only).  There were no cases of any dispatchers
that actually used the binding for anything except raising errors.
But it was necessary to make sure structural data didn't have broken
bindings if it wound up in error objects.

It was also left as an open question for the future as to whether or
not there might be some path dispatcher that did use the binding,
and could actually follow references to variables to their value.
But this hasn't come up, and it would also inhibit interesting
optimizations.

Since there is now a function for safely disconnecting relative values
from their specification (by connecting them to an on-the-fly
generated frame), it seems like a good time to change the picker
to a RELVAL.