Tweak bindings in path words to speed lookup
While the head binding of a path is used to get the binding process
started, linear searches for keys by identity in the context were
being used.  This was because the words could come from anywhere
(and be generated by GROUP!s).

With paths being seen by the user as immutable, even when aliased as
blocks, the idea of repurposing the bindings of the nested words to
point to the last place that word was found came up.  This may or
may not be useful, because an earlier thing in the path may have
changed to a different object.  But with the idea of using more
tuples and paths for specification of modules, having a consistent
chain on the left would happen frequently...and hence be needed
for optimization.

This makes that adjustment, cutting the time more than in half for
a simple test like:

    delta-time [val: 0 loop 1000000 [val: lib/add lib/subtract 1 2 3]]