Generalize "Arrow Words" and fix quoting/setting/etc.
Words containing `<` and `>` have presented historical problems, such
as not being able to assign them via `>:` or quote them with `'`.
There was also an open question about exactly how many forms of words
would be allowed, when it would come into conflict with TAG!

This commit patches in a self-contained bit of arrow-word code that
lets you build any WORD! you like out of `>`, `<`, `|`, `+`, `-`, `=`.
The only rule is that such words cannot both begin with a `<` and
end with a `>`.

A current special exemption is made for `<>`, which has a concept
for being both a tag and executable that has not yet been done.

Includes tests of transcode for the various forms.