Cleanup %base-infix.r using new SET-WORD! legality
Several workarounds were needed due to the inability of R3-Alpha to
create various symbolic ANY-WORD!s containing the < or > symbols.
Those workarounds are no longer needed, which tidies the boot process.

The ability to do this change without a new bootstrap executable is
due to work done to process files with PARSE instead of relying on
the bootstrap executable's LOAD facility.