Rename lambda from => to ->
The SHOVE operation was known to be a common and necessary thing, but
it was decided to be better as `>-`.

Lambda has subsumed POINTFREE and become extremely powerful.  It looks
lighter and cleaner as the plain arrow, which matches languages like
Haskell.  Though JavaScript uses => for lambdas, the fact that <=
is "less than or equal" suggests a non-arrow-like interpretation of
=> for consistency, as there's no emergent reason why they would
be diferent.