Eliminate usages of NEVERNULL() on REBVAL*
NEVERNULL() lets you annotate a pointer that says it's never null,
and gives it some teeth in the C build by stopping null assignments.
It uses deep voodoo so that it doesn't change how the C build does
assignments or references (e.g. it overloads -> and &).

It's a neat tool that came in handy for a debugging situation, and it
might do so again.  But because of the deep voodoo it doesn't combine
well with other more important tricks that want to use that voodoo.

This keeps the definition, but removes its usages on RELVAL*, which
are getting a more complex mechanism to check that REBCEL* are not
used in pointer math.