Stronger checking for REBCEL (const, no pointer math)
The REBCEL is a way of looking at a cell that sees it as fully
unquoted, even though it can be the same memory address as a quoted
cell.  The way it accomplishes its "view" is that you use different
routines for getting the type in "cell view" than when you are
treating it in ordinary "value view".  (The C++ build enforces which
functions you call via types, while the C build does not check as
they are the same datatype in that build.)

This adds two more useful protections.  One is that a REBCEL* is
always declared const, since it represents a view on what could
be shared data linked to by a QUOTED! cell that is shared among
other quoting levels.  The other is to disable pointer arithmetic,
like that which wuold be used in traversing an array--because a REBCEL*
might have been followed from a REB_QUOTED cell out to a single
array somewhere in memory, thus not inside the memory of the array
where it was being enumerated from.