Force PATH! access through VAL_PATH_AT() calls
In the historical implementation, paths were implemented underlyingly
the same way that BLOCK! and GROUP! were.  This meant that there was
always an allocated array node that could be extracted from them,
which contained all the items and knew its length.  The extracted
array could be passed to any generic routine that operated on arrays.

The new PATH! and TUPLE! concept is to avoid bad visual states where
they decay to appearing as single WORD! or INTEGER! by not letting
them be anywhere but the head position.  Also, the vision is that they
are immutable--so that they can never have fewer than two elements...
but this also permits freedom in implementing them in more compressed
ways.  That compression could pack all-integer tuples into a cell,
or turn a "refinement" style path like `/a` into a thin veneer over
a word cell, so that it took up no more space than in R3-Alpha.

In order to accomplish this, path manipulation operations can't expect
to extract an array out of a path cell, because they might not have
one.  Instead, they have to provide a temporary cell which may or
may not be used as the backing store for a fabricated cell which was
drawn out of the compressed representation...and use a VAL_PATH_AT()
accessor.