No error on inclusion of stdio.h in OS X 64-bit
There are ways stray printf() usage is avoided.  One is to catch if
`stdin` is defined before %sys-core.h, and the other is to refine any
usages of printf or fprintf via macro to a nonsensical name that
causes their usage to error.

Recent Mac OS seems to include stdio.h as a dependency from the
inclusion of basic non-stdio headers.  This switches to tolerate it
for all OS X 64-bit builds (which would be newer ones, as there is
no 32-bit build on newer platforms).