Fix PARSE on quoted series return result
The behavior PARSE was is experimenting with is to allow quoted
series, and make the result match the quote level of the input.  But
it had a bug if the input was quoted at a level higher than would fit
into a single cell, because it would copy the input cell and then
set the VAL_INDEX() of it.  However, this would be performing such
an operation on the REB_QUOTED cell.

This performs the necessary dequote/requote to do the operation safely.
The catching of this bug is thanks to the REBCEL(const*) change, which
enforced correct interpretation of the REBVAL* needing to run
CELL_KIND() if it were to match the type for mutable VAL_INDEX().