"blackhole" for opting into calculations, not results
Multiple return values used BLANK! in destination locations to say
that the caller is not interested in a particular output result.
Along with that concept came the idea that the called routine might
do less overall calculation.

However, this meant that sometimes functions that returned a result
would give you the result when you weren't actually interested in it.
You were faced with the choice of either opting out or opting in and
having to declare and pass in a variable.

This creates a middle ground using an empty ISSUE!, which visually
appears as a `#`.  This is referred to acting as "blackhole" and is
usable as the target of a SET to act as a no-op (illegal for BLANK!).
Because the blackhole is an ISSUE! it is truthy, making it a good fit
for the divergence in whether to run the calculation or not...and
because it is more rare than blank, it is less risky to have it
silently throw away a SET...and more likely to be intended.