EVALUATE as multi-return, expose invisible bit
Prior to multiple returns, EVAL experimented with using the new "@word"
types to abbreviate specifying a variable to return into:

     evaluate/set [1 + 2 3 + 4] 'var

     => evaluate @var [1 + 2 3 + 4]

While this tested skippable arguments and developed some new ideas that
were incorporated into "modal parameters", it was not as good as
multiple return values.  So this changes to use /RESULT as the
refinement name, and makes it an output:

     [pos value]: evaluate [1 + 2 3 + 4]

In addition, this changes the nature of the evaluator stepping to not
try to eliminate invisibles completely.  Instead, it makes them return
VOID! -but- annotates the return type for the position with a quote.
That makes skipping invisibles as easy as:

     until [not quoted? [pos value]: evaluate pos]

This adds a new capability to dialects using evaluate, in particular
to be able to literally examine slots both before and after invisible
locations.  Previously, anything after an invisible could only be
seen once it was evaluated.