Tune DELIMIT and PARSE handling of BLANK!, ISSUE!
This tweaks the handling of BLANK! to allow it to be used at "source
level" in dialects, while prescribing it as a synonym for NULL if
it is fetched from evaluation or a variable.

Being able to discern a "source" blank from a fetched one is thus
important, and the previous commit's tweak is used to move in steps
that allow blanks after comments to be discerned as source or not:

    >> spaced ["a" _ comment <b> _ "c"]
    == "a  c"

    >> spaced ["a" blank comment <b> blank "c"]
    == "a c"   ; acts as if you'd written `spaced ["a" "c"]`

DELIMIT is also modified to use tight spacing for ISSUE!, which is
an aspect of the transition to a merged ISSUE!/CHAR! datatype.  This
integrates well with the source-level spacing feature, for clear
results that don't require mixing SPACED and UNSPACED calls, and
thus can work cleanly with PRINT:

    project: 'Ren-C
    bad-thing: "Software Complexity"
    new?: does [project <> 'Rebol]

    print [#<< project #>> _ {The} (if new? 'NEW) {War On} bad-thing]

    str = "<<Ren-C>> The NEW War On Software Complexity"

This also changes PARSE to embrace the feature of source-level BLANK!
representing a synonym for SKIP:

    >> did parse [1 <two> "three"] [integer! _ text!]
    == #[true]

While fetched blanks act as NULL-equivalents:

    >> subrule: _

    >> did parse [1 <two>] [integer! subrule tag!]
    == #[true]

It was actually the case that PARSE was not allowing NULL parse rules
when it should (due to not being sync'd with the void-as-misspelled
change).  So routines like REWORD were using BLANK! for nulling out
rules made via reduce.  Changing to NULL was the easy fix, but the
usage of REDUCE made it fail in a tricky way...so this provided the
critical momentum for REDUCE to mutate NULL into VOID! by default.
Predicates will allow usage of `reduce .try [...]` to get blanks from
nulls, or `reduce . [...]` to use the identity function and dissolve
the nulls.