Allow all "taglike" arrow WORD!s, some tag scan cleanup
This rationalizes the embrace of `<>` as a WORD! by saying that all
taglike-looking-things which are built up of only the arrow word
characters `<`, `>`, `-`, `=`, `+`, and `|` are WORD!s.  That permits
the "spaceship operator" `<=>` as well as things like `<->` or
`<+>` which could be used for creative purposes.

Some dead code related to handling special exceptions prior to the
arrow words rule is deleted, and errors are now raised on cases like
`>abc<` or `<abc>def`, thus requiring tags to be up against delimiting
() or [], or spaces/newlines.