Change scanning to not consider `.` a WORD-character
The historical scanner considered `/` to be a delimiter-character,
which would cause token scanning to stop so that individual tokens
could be collected one at a time into the PATH!.  Data types that
had exceptions and did not consider slash to be a delimiter needed
special behavior to advance the "pre-scanning" end position...which
is what FILE! and URL! had to do.

With generic TUPLE!, the handling for `.` becomes more like `/`.  It
becomes a delimiter character, and can no longer be put in WORD!s...
which is by design.

However, some data types like EMAIL!, TIME!, and MONEY! were using the
non-delimiter status of `.` to accept the end pointer from the basic
prescanning process.  This hacks in an exception that lets them
subsume one or all dots before a non-dot delimiter is seen.