Abstract Make "Fake Type Constraints" Mechanism
Consolidating types like REFINEMENT! into a specific instance of PATH!,
or CHAR! into a specific instance of ISSUE!, reduces the total number
of fundamental datatypes and offers consistency and flexibility.  But
there is no standard mechanism for expressing a type constraint in a
function spec (e.g. "integer!, but it must be even") so the unification
causes a loss of that check.

A true solution to the problem needs to be found.  But until it is,
this takes a compatibility hack and makes it clearer in the code what
it is doing.  It creates some fake values that can be used by function
specs which at least give an annotation of the constraint.  They are
in Lib_Context so that native specs can use them.

While they have no teeth in typeset creation (they only verify that the
unconstrained form of the type matches), PARSE and MATCH recognize the
symbol and enforce it.