Unify ANY-TUPLE! and ANY-PATH! to ANY-SEQUENCE!
This is a coarse unification that can boot and read https sites, but
is otherwise very much a work in progress.  The nuances of byte-based
behavior will need to be folded in so that they work on paths, and
have appropriate error conditions for when integer preconditions are
not met (e.g. to divide tuples like 'a.b / 'c.1)