PATH!/TUPLE! precedence hacks and begin test file
An idea behind the path and tuple scanning is that paths have the
"lower precedence", hence `a/b.c/d` is a 3-element pah with a tuple
in the middle, and `a.b/c.d` is a 2-element path with two tuples.

Because of the ad-hoc nature of the scanner, the additions of new
types and scanning in Ren-C has also been hacked in--and is in need
of a more spec-driven rewrite.  However, this tweaks the scanner to
give the desired results for a few more cases.

More importantly: this starts a file that does structural tests on the
scan results.  The tests use BLOCK! and GROUP! as proxies for PATH! and
TUPLE! in the loaded data, so that mistakes in transcoding a string
are caught vs. comparing to LOADed data which would have the same
invalid results.  The structural checking is able to test things that
could mold ambiguously.