Make leading `@` and `:` handled in separate step
The mechanism of R3-Alpha for making GET-PATH!s was to scan a GET-WORD!
and then when forming the path, notice it had a GET-WORD! in the
leading position and converted it.  This technique could not work
in Ren-C for BLANK!-headed paths (there's no GET-BLANK!) so a hack
was put in place to set a bit on that BLANK!.

This replaces that hack with a more general method of setting a
"pending sigil" which will apply the SYM-ification or GET-ification
after the whole token is completed.