Simplify handling of out-of-turn dot or slash
This changes the behavior on seeing an out-of-turn dot or slash token
to simply pushing a blank, and falling through to all the work for
sorting out the cases to be done in the main "lookahead" step.

As a result, the scan tests that have been written so far for nesting
paths and tuples together all pass.