Add lazy "rebUnrelativize" for error handling
Generally speaking, things like errors attempt to keep the bindings
for any words that are put into the error.  Unrelativize() was added
for cases where threading a specifier would be prohibitive.  This
makes an easier way to call Unrelativize() without a DECLARE_LOCAL(),
by putting the value into an allocated API handle.  If the RELVAL*
is being unrelativized to pass to an error, the handle will be
freed automatically.

This is useful in several cases being caught in the FFI which were
passing relative values to fail(), caught by the added type checking
in the C++ build.