Implement Sequence Compression for byte-sized ints
This brings the feature of fitting up to 8 byte-sized integers into
a cell on 32-bit platforms, and up to 16 byte-sized integers into
a cell on 64-bit platforms.  So the immediate value that is TUPLE! is
not any bigger than in R3-Alpha, and this applies to ANY-PATH! also.

Currently the optimization does not apply to making sequences from
already-existing arrays...only from scanning.  Making from arrays is
TBD--which might imply not passing in already created arrays, but
rather pushing to the stack or giving a pointer into the array data
to be copied.