Formalize "Sequence" Name, Organize Files
The term "Sequence" was used to mean "non-array-series" inside of the
code, to help call attention to when there was a special routine that
had to be called for arrays for a given operation.

Need for the name for the collective TUPLE! and PATH! types is higher
priority than that distinction.  So this introduces a header file with
a bit more organization as the feature comes together, called
%sys-sequence.h

The pathing evaluative features are thus moved into %sys-pick.h to
cover path picking and poking, although it could be called "sequence
evaluation" or something like that.

The %sys-tuple.h file is folded into the sequence header, with its
messy residual routines getting slightly reigned in with the right
terminology for sequences.