Use underlying "heart" byte REB_ARRAY for PATH!
(Booting but hundreds of tests still failing, hence still a WIP.)

This builds on the distinction of the fourth byte in the cell from
the second "KIND_BYTE()" for PATH!s that are implemented using arrays.
It formalizes the idea that REB_PATH is always implementation-wise
done with a representation dicated by the fourth byte, which is never
an ANY-PATH!.

Rather than call the fourth byte MIRROR_BYTE() this begins a change
to calling it the HEART_BYTE().  For the moment they are set up as
synonyms to show code that has put some consideration into the
distinction vs. not...and MIRROR_BYTE() should be slowly phased out.