Patch contention between SET-PATH! and TIME!
The new mechanism for establishing SET-XXX! is generic, based on
a shared lookahead at the end of token scanning for a `/` or a `:`
that is attached without spacing.

There were some forms of time that were allowed to end or begin with
a `:` that were causing ambiguity.  A workaround was established for
that, but this extends the workaround to cover more tests.

(Note: The scanner really needs a complete redesign when things are
more finalized.)