Add Tuple to ANY-SCALAR!, test fixes
TUPLE!'s rethinking now has it acting as more than a scalar, but
several functions use ANY-SCALAR! as a type constraint which expect
to allow tuples.  In theory, the code could now be applied to PATH!
or GET-PATH! or SET-TUPLE! etc.

For the moment, this just adds TUPLE! back to ANY-SCALAR! for the
purposes of cutting down on test failures, until a better idea for
dealing with this comes along.