Version in %make-boot.r immutability fix
The file %version.r contains a TUPLE! that historically has been
loaded and patched with additional information from the specific
build being done.

This patching was done by appearing to "mutate" the tuple.  However,
since tuples are immutable what was actually happening is that
`version/3: 10` actually assigned a new tuple to version with a
different third value.

Such a feature is misleading, because it's not changing the value
as it would be seen by "other references".  Whether this form of
variable update should be legal or not can be debated.

This instead makes a new tuple via TO TUPLE! and REDUCE with the
supplemented values mixed with the original values.