Use MAKE-FILE in bootstrap
This starts the process of getting rid of the implicit pathing, e.g.

    >> dir: %some/where/

    >> dir/stuff/tacked/on
    == %some/where/stuff/tacked/on

The handling of slashes in this code was sketchy (tolerated having or
ot having), and it was a bad citizen in the attempts to come up with
new designs for PATH!-ing.  The MAKE-FILE dialect aspires to replace
this method.

This goes ahead and uses it in the rebmake files.  But due to bootstrap
issues it only uses the BLOCK! form of the dialect.  Also, the buggy
parsing of "NewPath" styled paths like `foo/` in the bootstrap build
mean it spells out directories as `foo /`.