Remove superfluous VAL_ACT_XXX accessors
The system has attempted to make it possible to pass around an ACTION!
and have it be fully implied by that single handle.  This means that
value-oriented accessors are not essential, but merely conveniences.

In light of a new desire to make value cells cache the name that
particular assignmet was done through, the value-based accessors are
a distraction from that feature.  Also, the names VAL_ACT_XXX do a
contraction that has not been followed in oter VAL_XXX routines,
e.g. VAL_ARRAY or VAL_SEQUENCE_AT.