Labels for ACTION!s fetched from GET-WORD!/GET-PATH!
This is a trick which stores the symbol an ACTION! was last fetched
from into its cell.  The space it occupies is what would be the
"details" in archetypal functions (e.g. those living in a paramlist).
But the way this works is to check the bit in the node to see if it
is an array series...and if it is then it assumes it is an archetype
(no cached name) otherwise it should be a REBSTR* node, as a label.

    >> do compose [(:append) 3 <d>]
    ** Script Error: append does not allow #[datatype! integer!]
       for its series argument

Function compositions that use intermediate variables to store their
work as they go currently contaminate the name, so if a composition
is not stored in a word before use it will lose the name.  This
would require special handling.