Reserve .5 and 5. for now
The concession to represent two-element INTEGER! tuples e.g. 5x5 is
useful, because it allows the PAIR! type to be folded into TUPLE!.

However, the blank-headed tuple [_ 5] and blank-tailed tuple [5 _]
have no easy alernate representation.  `x5` can't be used because
that collides with variable names (and it also looks bad).

DECIMAL! has an easy workaround to just always have 0 in these spots.
That would free the representations .5 and 5. for TUPLE!.

The generality would be hurt to disallow these TUPLE! representations,
which are otherwise mechanically legitimate and parallel to the
PATH! forms /5 and 5/ - it seems greedy of DECIMAL! users to demand
it.  However, since generic tuples are new and their usefulness in
being general is not fully appreciated...this doesn't make any
decision, but just reserves the representations for future use.