Improve Init_Action() to account for labeling
This commit forces each action value initialization to be conscious of
what label it is using.  An alias for a null REBSTR* of "ANONYMOUS"
is introduced to call attention to when an action would have no name
in the debug trace, encouraging the search for a better thing to pass.

It also shifts the generation of derived functions via SPECIALIZE
etc. to use GET to fetch actions out of words...because the idea is
that GET will not change the name of the action to match that of
the name in the fetched location.  Whether this is the best idea or
not, it seems to work for the moment:

    >> do compose [(specialize 'append [value: <d>]) 1]
    ** Script Error: append does not allow #[datatype! integer!]
        for its series argument