Cache action symbol in non-archetypal FRAME! values
This expands the trick used to put symbols into ACTION! value cells
when they are non-archetypal to FRAME!s.  Then, when the frame is
used in an executable context (like DO FRAME!) or becomes an ACTION!
again (via `make action!`) the tunneled label is carried through.

Stack labeling and errors benefit as a result:

    >> f: make frame! :append
    >> f/series: 1
    >> f/value: <d>

    >> do f
    ** Script Error: append does not allow #[datatype! integer!]
        for its series argument