Change POINTFREE variadic form to <- instead of ->
For perceived convenience, the lambda generator `->` was unified with
a variadic form of POINTFREE that would choose behavior whether the
thing to the right of the arrow was a BLOCK! or not.

This made the much "simpler" lambda have to be variadic, complicating
its usage.  It also ran the risk of people inadvertently getting
POINTFREE behavior just by leaving off a BLOCK!.

Given that -> is chosen for a function generator, the idea that <-
does anything that isn't tied to being a function generator would be
kind of confusing.  So this makes the move of putting the variadic
pointfree behavior onto <-.  That means that all lambdas can be
assured of having a block on the right hand side.

(cherry picked from commit 4057d595868963ddc9db6c85da5e0d8c547a5ae3)