Eliminate "stackful" Debug Hook Concept
An idea for generating a completely "hookable" evaluator was to make
the evaluator function a pointer that could be replaced.  Writing
a hook would involve replacing this with your own function that got
a chance to see each frame and could do some inspection or transform
prior to calling the main evaluator function (if it chose to).  It
could also do some inspection of the results afterward.

This "ENCLOSE"-style hook for the evaluator itself had a few problems,
but the main one is that it inhibits stackless processing.  This way
of using the C stack to wedge control over the evaluator inhibits
unwinding across the frames.  With some significant work done on a
stackless branch of the code, there will be no further following with
this line of thinking.

Deleting the code makes more sense than keeping it for reference in
writing a new debugger, as the code is available in git if it needs
to be referred to.