Kill SHORTHAND(), kind.byte optimization
The evaluator needs to frequently access nested frame fields, and this
needed aliases.  Instead of using macros to do it (making assumptions
about the variable name into which nested access was being done), a
SHORTHAND() method was introduced that aliased the variable.

Since it turns out that the variable is almost always called "f", this
switches over to using macros...which are easier to understand, and
won't incur overhead even in debug builds.

Also: there was at one point a notable slowdown in debug builds
from excessive checking of the cell's validity during VAL_TYPE().
This later was mitigated somewhat by the usage of KIND_BYTE(), but
even that was a bit slow compared to KIND_BYTE_UNCHECKED().

In addition to noting that KIND_BYTE_UNCHECKED(v) was long and awkward
to type, sometimes there were cases where the kind of the value was
needed to be out of sync with that of the current value...but then
could be sync'd up to the value.  This encouraged attempts to "beat
the optimizer" and manually maintain the kind byte state when it could
be intentionally out of sync...running the risk of being intentionally
out of sync in the process.

With minor rewriting and use of the new #define macro shorthand
convention, this rethinks the issue.  The two times that the kind.byte
got out of sync are addressed by getting the kind from another place
it is available at that moment (the f->out), and using a switch()
statement.  No noticeable performance degradation results.