Split %c-eval.c into two files, actions get own frames
This change breaks the evaluator into two main parts: one part that
walks a block and handles items like WORD!/PATH!/SET-WORD!/GROUP!/etc.
The other part handles the dispatch of actions.

One aspect of this change is giving actions their own independent
REBFRM*, and hence making it theoretically possible to have reified
FRAME! entities representing the concept of walking a block.  This was
avoided initially for efficiency reasons, but work on the stackless
build and trying to make a generic debugger design meant there had
to be a way to refer to a block evaluation persistently from usermode.

Because this slows things down--especially in the debug build--further
commits will focus on trying to get the cost down.