Adapt REBFRM* to be allocated from memory pools
This makes the REBFRM* layout compatible with the memory pools, by
enforcing the bit structure in header bits that they require.  This
also broke the means by which REBFRM* appeared different from
paramlists and contexts by having the NODE_FLAG_CELL bit set...so
it had to be adjusted to keep that discernment working.

Several callsites had not been using EVAL_MASK_DEFAULT, and this
discovered those in the process.