Eliminate specialized Prep_Cell() of f_arg
In order to avoid the GC process seeing uninitialized bits, R3-Alpha
would fill the argument and refinement slots for a function call with
NONE! values.  One early idea in Ren-C that was conceived to help with
both performance and correctness would be to skip this step, instead
making the GC aware of how far the evaluator had gotten in the frame
with argument evaluation...only marking the filled arguments.  This
allowed "partially filled" frames to be identified more clearly as
containing trash, so that they would not be leaked to the unaware user
despite their partialness.

This idea became more complicated when stack cells were differentiated
from heap cells with different forms of "format bits".  That concept
never quite squared with the idea that it would be the evaluator
filling in the bits...since it didn't really know whether it was making
a stack frame or a heap frame.

The idea was in flux for a while, but the concept of formatting cells
as having "stack lifetime" ceased to be viable with the stackless
implementation.  With that idea gone, a separate identified step of
having to test whether the `f->arg` should be initialized or not can
be avoided, instead just letting the preparation happen at the moment
that the argument is fulfilled with its value.