Use full space on 64-bit for tuple packing
The existing code for tuple using a byte buffer was built on the
"common" union field of the byte payload of a cell.  This field is
restricted to the common binary size between 32-bit and 64-bit.

This switches to using the `varies` field, which means that larger
tuples will avoid allocations on 64-bit platforms which would need
them on 32-bit platforms.

Also fixes a problem found by R3_ALWAYS_MALLOC with some of the
"dying" tuple operation code that is being kept alive for tests, which
isn't geared for tuples of sizes > "MAX_TUPLE_LEN".