Introduce STATE_BYTE() for frames
When frame flags were initially introduced, the platform-sensitive
byte ordering tricks for implicit termination had just been done.
This made it seem like it could be interesting to have the frame
flags implicitly terminate the "spare" cell...the same way that it
terminated singulars.

It turned out that wasn't used all that often (and in fact, there's
some question as to whether the termination requirement for arrays is
even a good idea...considering that the terminators have to be kept
in sync, and force allocations of array sizes larger than the content
to account for a terminator).

With the frame transitioned to being a memory pooled allocation, the
flags are no longer in a spot that terminates a cell...so that frees
up the 8 bits that were being forced to zero for the terminator.
Rather than just use those for flags--however--this groups them to
act as a "state byte".  This choice is based on experience with the
stackless branch, which makes heavy use of this byte.

Applying it to manage the "state" of the main evaluator allows it to
more clearly jump to the reevaluation and lookahead steps.  It also
frees up two flags that were taken for implementing the more
circuitious approach to doing the same thing.