Make NO_LOOKAHEAD apply to quoted arg fulfillment
When enfix arguments were being fulfilled evaluatively, e.g.
`1 + 2 * 3`, there is specific code inhibiting the gathering of the
right hand `2` argument from seeing the `*` while the `+` is still
running.

This exemption was not applied to arguments that were soft quoted
(even if they evaluated) or hard quoted.  But this goes ahead and
makes it applicable to all the cases, as not being so was breaking
the handling of `if false [1] then x -> [2] else [3]`.

A note is added to the evaluator about the forum post explaining some
of the details of that composition.