Adjust C standard compliance check for TO-C-NAME
The TO-C-NAME function had a well-meaning feature to warn you if you
were creating a name reserved by the C language standard.  This would
be names starting with an underscore under certain circumstances.

The check seems to have atrophied due to inconsistent use of ISSUE!
vs. WORD! to indicate the desired scope rule.  Once it was fixed, it
pointed out that the check wasn't accounting for prefixes which may
make it okay to use TO-C-NAME instances as a suffix that would be
invalid if they were used standalone.

This changes CSCAPE to just assume you're using the prefixed form
and it's okay.  So only non-CSCAPE calls are really checked.  It's
still better than nothing, and may at least serve as a reminder to
look over the code for potential non-standards-compliant names.