EXTRA/PAYLOAD bytes rename, REB_BYTES sequence
(Note: This commit adds a type to %types.r that needs to reclaim
CHAR! for the bytes type, hence it requires the next commit)

Using the same name (`common` and `varies`) for the PAYLOAD raw bytes
structure and EXTRA structure led to a bug, where the wrong sizeof()
was used.

This corrects the bug, and tries to avoid it in the future by improving
and differentiating the names of the fields.  The new names say how
many bytes are implied (at_least_X or exactly_X)...which has the
benefit of differing between the EXTRA and PAYLOAD where X is 4 and 8
respectively.

The commit also updates the Bytes-using ANY-SEQUENCE! implementation
to use a new "internal" heart byte called REB_BYTES.  The previous
choice of REB_CHAR was simply to pick something that didn't have any
GC behavior with nodes, but was (a) misleading and (b) this no longer
can be used in light of the coming ISSUE! and CHAR! unification.
Using a distinct dedicated type clarifies what's going on.