Make the VAL_UTF8_AT() generic forms more explicit
Both ANY_WORD! and ANY_STRING! can have UTF8 extracted from them,
so there was a generic extractor besides VAL_STRING_AT().

This makes the naming for that generic extractor more systematic, and
switches to using it in more places as the default way for getting
a const UTF-8 data pointer from a value.  This makes code at the
callsites clearer, generally breifer, and easier to change when new
ideas come up.  It also paves the way for being polymorphic with a
UTF-8-bearing cell that stores its data in the cell payload.