Make it easier to add non-types to %types.r
There are some usages of the type (or "Kind") bytes which do not
actually correspond to a DATATYPE! in the language.  REB_NULL and
REB_0_END have been examples of this, but others are have been
maintained beyond REB_MAX and documented in the %make-boot.r file
in the enumerated type, as opposed to in %types.r.

This is a start on making it easier to put non-types into %types.r;
they are flagged with the name being an ISSUE!.

The internal constant REB_NULLED becomes REB_NULL when the names are
systemized in this way, but that is not such a bad consequence...as the
REB_ prefix helps know it is talking about something Rebol-specific vs.
C NULL.  The rest of potential confusion is hadled by Init_Nulled()
being named thus instead of Init_Null().