Get rid of separate %base-infix.r file
R3-Alpha was less permissive about allowing things like `<:`, so a file
called %base-infix.r of the word forms like `<` was specially handled
by the boot process to load them and set them up as infix symbols.

That purpose ceased to be applicable, but the file stuck around as
a central place to look at all the infix definitions together as
a group.

Having them grouped together in this way didn't necessarily fit the
ordering definition of the symbols...as some were higher level and
some were lower level functions.  The arbitrary positioning of *all*
the infix symbols defined in one file meant that useful things like
`=` were being defined too late to be used in low level primitives,
which had to use `equal?` instead.

Given that order doesn't make any sense, this disbands the file and
empowers base definitions to use `=` and other operators.