Use AS TEXT! instead of TO WORD! for REB_0_END
The mechanisms that were designed to prevent word conversion of `0-end`
actually worked.  But this prevented bootstrap, because it was catcing
the creation of a bitflag which went through a WORD!-converting step
with `0-end`.

This switches to using AS TEXT!, since TO TEXT! on issues in the
bootstrap build keeps the leading `#`.