Rename MIRROR_BYTE as HEART_BYTE
The concept of having cells contain two different bytes for their
type originated with the concept of a container--e.g. a BLOCK!--which
would have a KIND_BYTE of REB_BLOCK and a MIRROR_BYTE of REB_WORD.
If the mirror was "broken", that meant the cell was actually for a
WORD!, and the block contained a WORD!.  The idea would be that you
could not store a single-element BLOCK! in a BLOCK! in this way, hence
if the type matched it was supposed to be a "mirror".

The word is confusing...as what the byte really evolved to mean is
"the byte that the GC pays attention to"...e.g. the implementation
bits.  An easy, short, and not-yet-taken name to use for this is the
HEART_BYTE, so this switches over to using that.

(Unfortunately, the idea of generically using the HEART_BYTE to wrap any
item in a BLOCK! faces the challenge that array enumeration currently
expects all arrays to terminate with an end marker.  However, the
byte has been used to allow the implementation to vary for things like
ANY-SEQUENCE! and ISSUE!, to fit data into a cell without requiring
an additional allocation.)