Cut down on usages of SPECIFIC()
The SPECIFIC() inline function tests a RELVAL* to make sure it does
not contain a relative binding before promoting it to a REBVAL*.

This can be a useful test, but not particularly so when the bits were
just initialized and you can tell that the value is specific.  (e.g.
if you've just initialized an input RELVAL* cell with the bits of
a non-bindable type, like an integer).

Using SPECIFIC(v) in places that a cast(REBVAL*, v) would probably
suffice was taking over 5% of the execution time in debug builds.  This
commit cuts the SPECIFIC() calls back a bit...making them about 1/10
the previous cost.