Don't use T_String in T_Issue, SIZE reflector
When an ISSUE! wasn't "char!-like" it was deferring to the string
type to answer questions like LENGTH OF, or for generics like COPY.
However, that deferring was being done for longer sequences that still
fit in cells...which the string dispatcher wouldn't understand.

This goes ahead and implements the generic LENGTH and COPY operations
inside of ISSUE!.  It also adds a SIZE reflector both the ISSUE! and
ANY-STRING!, to get a byte-count for the binary encoding of UTF-8
without having to do a binary conversion to get it.

It's not clear when--if ever--deferring to the ANY-STRING! dispatcher
would be appropriate for ISSUE!, since all operations would have to
be implemented in the ISSUE! reflector for shorter strings anyway.
Review the general concept.