JOIN and REJOIN work, includes enabling for ISSUE!
Historically, JOIN is based on APPEND; and immutable types have raised
questions on the right way to implement it.  Additionally, the REJOIN
interface is historically somewhat unpredictable.

This commit includes a little bit of modernization of the JOIN and
REJOIN usermode functions...to try and deal correctly with things like
invisibles and the NULL/BLANK distinction.  In the process, it adds
support for ISSUE! for both JOIN and REJOIN.