Only use REBCEL(const*) checks in DEBUG_CHECK_CASTS
Most checks for REBCEL() can be done with just a derived class in the
C++ build.  However, the specific check that is used to avoid pointer
math on the cells themselves requires a heavier wrapper and override
of the `->` operator.  This reserves that for DEBUG_CHECK_CASTS builds,
as otherwise it can cost ~ 10% of the runtime.