Use EVIL_MACRO method for TRACK_CELL_IF_DEBUG
Due to the debug build's lack of inlining, there's added overhead for
calling functions on basics inside of basics like RESET_CELL.

This changes the initialization of the tracking info to use an "evil
macro" which repeats a parameter.  This would be bad if it were
evaluative...but it's not.

Interestingly, the process actually makes it clearer what each tracking
mode is doing.