Require space before `;` is considered a comment
Historically, Rebol allowed semicolons to abut directly against
other tokens in order to start a comment:

   rebol2>> load "a;b"
   == a

This was relaxed in Ren-C for URL!, given that semicolon is a valid
character in urls:

   >> load "http://example.com/a;b.html"
   == http://example.com/a;b.html

With the introduction of ISSUE!, there is now another type that could
benefit from allowing semicolons that are not "spaced out" to merge
into the token itself:

    >> if #; = second "a;b" [print "Convenient."]
    Convenient.

However, an policy of "sometimes merging with the token and sometimes
not" seems shaky.

Given the overall theme of space significance in the lexical model,
this makes the shift to where semicolons are only considered comments
when they are spaced out.  Token types that don't accept semicolon
will now raise errors:

    >> load "a;b"
    ** Syntax Error: invalid "word" -- "a;b"

Empirically it does not seem like the idea of such "tight" comments
was very popular (and the Ren-C coding standards prescribe *TWO*
spaces off of the code for line comments).