Split set ops from bitwise e.g. NOT+ and COMPLEMENT
R3-Alpha had some mixture of the concepts of "set operations" like
INTERSECT and UNION with the bitwise operators like AND and OR.

This mixture glossed over the cases where these could be different
operations, e.g. BINARY!:

   r3-alpha>> intersect #{0102} #{0203}
   == #{02}  ; INTERSECT treates as collection of bytes

   r3-alpha>> #{0102} and #{0203}
   == #{0002}  ; AND treats as two 16-bit numbers

The set-based operations also had refinements for /SKIP and /CASE,
which are considerations irrelvant to bitwise math...further pointing
to the difference.

For a time, Ren-C pushed greater unification of these operations
based on the proposal to make operations like AND and OR be conditional
and treat their operands based on conditional logic, returning only
byte-collection INTERSECT demonstrated above.

Current thinking is that these categories of operations should be
kept separate.