Get rid of separate Init path for ANY-STRING!
The `Init_Any_String_At_Core()` function was distinct from the
`Init_Any_Series_At_Core()`, which made it hard to write generic
functions for ANY-SERIES! initialization.  This was an artifact of
some transitional-phase code during the UTF-8 Everywhere conversion,
and doesn't make sense to keep around now.

This commit puts all series initialization through the common
routine, and also makes the array and string initialization variations
do compile-time type checking on REBARR* and REBSTR* respectively.

Additionally it moves the core series initialization routine into
the header file so that it can be inlined, as the release build code
is very short.