Remove non-working test idiom from LOAD
As an example of how to handle the reuse of the name ALL in a
refinement, a test of creating a context in which ALL would be
undefined was created and then bound with `<in>`.  In practice, this
was not working...as the local still overrode at the time of function
definition.  The idiom is still important to think through, but
will need a better mechanism.