Refactor modal arguments to work with enfix
The code for handling modal arguments was processed using a GOTO which
did not allow a variance between the handling in normal processing and
enfix processing.  This allowed flags to get out of sync and trigger
an assert when used as infix.

This factors the code out into a common function, so the callsites
can manage the flags distinctly as necessary.