Rethink AND/OR/XOR as non-short-circuit by default
The original pitch for changing AND/OR/XOR to be "conditional" and
not bitwise (as NOT is) was that it would match the expectations of
those coming from other languages.

When Ren-C tried this concept out in practice, it was observed that
another expected property from most languages is short-circuiting.
However, Rebol is not built to do this unless the code is put in
some kind of container (e.g. a BLOCK!) which can hold it for an
evaluation that may or may not happen.

Trying to appease the need for this conditionality, it was experimented
with taking branches on the right...and using quoting semantics to
catch accidents that were written in a non-branching style.  This
led to unpleasing aesthetics:

    if x and [y] [
        print "The block on the right is confusing"
    ]

This does not have the desired improvement to flow over ALL [...],
especially when the option of ALL...THEN exists:

    all [x | y] then [
        print "This feels cleaner"
    ]

Having experimented with the short-circuiting ANDs and ORs for a while,
they weren't very pleasing.  So it seems that a non short circuiting
form by default isn't such a bad thing.  It is likely more efficient
to fetch most variables than to start a new frame, so there's not
really anything wrong with `if x and y [...]` if that's the case.

This commit also changes the result to return a LOGIC! and not try and
return the truthy value itself (as ANY and ALL do).  That was
motivated by the desire to use AND and OR as variations of THEN and
ELSE that were triggered by logic instead of by nullity.  With lambdas
that can generally be achieved in a more readable fashion:

    ... then flag -> [
       if flag [...]
    ]

(Also, coming changes to branches should enable null-returning branches
to be marked as such, which reduces the need for this.)

HOWEVER...in case short-circuit forms want to be used without inventing
a new word for them, the /SHORT refinement is added to AND and OR.
This lets them take a branch as an argument.  It is a modally triggered
refinement, so putting an `@` on the branch causes it to apply

    >> false and @[print "running!" true]
    == #[false]

    >> true and @[print "running!" true]
    running
    == #[true]

It is a feature people may be interested in using or not--but the key
is that short-circuiting is no longer forced on people by default.