Merge BINARY! and ANY-STRING!/ISSUE! find routines
The combinatorics of searching in binaries and strings was already
somewhat fractured in R3-Alpha...where /SKIP counts were not handled
in binary searches, and there were various broken edge cases.  This
got a bit more jumbled with the introduction of UTF-8 everywhere, as
some fledgling routines tried to make it possible to do searches for
binaries in strings and vice-versa.

This commit lays down the law that you cannot search for an arbitrary
BINARY! in an ANY-STRING!, due to the fact that such a search could
land on "half a codepoint" and not be frameable in terms of the FIND
or PARSE return results.  Searching for ANY-STRING! in BINARY! is legal
however, as strings are subsets of binaries and the binary index
affords a return at any offset within the binary.

Because handling of things like negative skips and the various edge
conditions is difficult to get right, this folds "binstr" searches
into one routine that handles searching for one binary or string in
another.  This generalization means passing in cells instead of
passing in series, which means some callsites may need to fabricate a
cell to do so...but that is seeming a more acceptable practice for
these cases than trying to pass the parameters separately.

Optimizing this routine--e.g. for caseless searches with skip of 1
to use memcmp()--will be important.  But having a solid generic
routine is necessary, especially if it's going to take into account
"proper" unicode case folding.